// src/api/schema.ts
export interface components {
    schemas: {
// ...
        ProjectResponse: {
            /**
             * @description Уникальный идентификатор проекта
             * @example 10
             */
            id: number;
            /**
             * @description Название проекта
             * @example Capacity Planning
             */
            name: string;
            /**
             * @description Описание проекта
             * @example Сервис для планирования загрузки студенческой команды
             */
            description?: string | null;
            /**
             * Format: date-time
             * @description Дата создания проекта
             * @example 2026-03-13T10:00:00Z
             */
            created_at: string;
            /**
             * @description ID владельца проекта
             * @example 1
             */
            owner_student_id: number;
        };
// ...
        ProjectListResponse: {
            items: components["schemas"]["ProjectResponse"][];
            /** @example 25 */
            total: number;
            /** @example 1 */
            page: number;
            /** @example 20 */
            page_size: number;
        };
// ...
    };
// ...
}